+++ src/source/source_cache.test.ts
@@ -759,12 +759,14 @@ describe('SourceCache#update', () => {
         sourceCache.onAdd(undefined);
     });
 
-    test('retains children for fading when tile.fadeEndTime is not set', done => {
+    test('retain children for fading fadeEndTime is 0 (added but registerFadeDuration() is not called yet)', done => {
         const transform = new Transform();
         transform.resize(511, 511);
         transform.zoom = 1;
 
         const sourceCache = createSourceCache({
+
+            // not setting fadeEndTime because class Tile default is 0, and need to be tested
             loadTile(tile, callback) {
                 tile.timeAdded = Date.now();
                 tile.state = 'loaded';
@@ -782,36 +784,6 @@ describe('SourceCache#update', () => {
                 sourceCache.update(transform);
 
                 expect(sourceCache.getRenderableIds()).toHaveLength(5);
-                done();
-            }
-        });
-        sourceCache.onAdd(undefined);
-    });
-
-    test('does not retain children for fading when tile.fadeEndTime is 0', done => {
-        const transform = new Transform();
-        transform.resize(511, 511);
-        transform.zoom = 1;
-
-        const sourceCache = createSourceCache({
-            loadTile(tile, callback) {
-                tile.timeAdded = Date.now();
-                tile.state = 'loaded';
-                tile.fadeEndTime = 0;
-                callback();
-            }
-        });
-
-        (sourceCache._source as any).type = 'raster';
-
-        sourceCache.on('data', (e) => {
-            if (e.sourceDataType === 'metadata') {
-                sourceCache.update(transform);
-
-                transform.zoom = 0;
-                sourceCache.update(transform);
-
-                expect(sourceCache.getRenderableIds()).toHaveLength(1);
                 done();
             }
         });